rename: 테스트 파일 productList 로 경로 변경
